+++ src/ContaoArtikeldatumBundle.php
@@ -1,5 +1,14 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * Dieses Bundle stellt das Inserttag {{article_update}} bereit.
+ *
+ * @author    Frank Binding
+ * @license   LGPL-3.0-or-later
+ */
+
 namespace Schachbulle\ContaoArtikeldatumBundle;
 
 use Symfony\Component\HttpKernel\Bundle\Bundle;