+++ src/ContaoManager/Plugin.php
@@ -1,5 +1,14 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * Dieses Bundle stellt das Inserttag {{article_update}} bereit.
+ *
+ * @author    Frank Binding
+ * @license   LGPL-3.0-or-later
+ */
+
 namespace Schachbulle\ContaoArtikeldatumBundle\ContaoManager;
 
 use Contao\CoreBundle\ContaoCoreBundle;
@@ -10,14 +19,14 @@ use Schachbulle\ContaoArtikeldatumBundle\ContaoArtikeldatumBundle;
 
 class Plugin implements BundlePluginInterface
 {
-	/**
-	 * {@inheritdoc}
-	 */
-	public function getBundles(ParserInterface $parser)
-	{
-		return [
-			BundleConfig::create(ContaoArtikeldatumBundle::class)
-				->setLoadAfter([ContaoCoreBundle::class]),
-		];
-	}
+    /**
+     * {@inheritdoc}
+     */
+    public function getBundles(ParserInterface $parser): array
+    {
+        return [
+            BundleConfig::create(ContaoArtikeldatumBundle::class)
+                ->setLoadAfter([ContaoCoreBundle::class]),
+        ];
+    }
 }